Fix: Distance metric not enabling clustering
The attractive forces of similar units weren't strong enough compared to the repulsive forces of dissimilar units. Nodes ended up forming a circle at the edge of the maximum node range.

This was fixed by making the distance between similar nodes exponentially smaller the more similar they are. Some fine tuning will be needed in the future to improve the node layout.