fix: Incomplete edge distance calculations
This was caused by the fact that the same iterator was being used for two loops. Iterators are not reset after being exhausted by a loop, so this nested loop was only being used once. This caused one node to form an edge with all nodes, but no other edge being formed between the other nodes.